refactor(marking): inline goals assessment into marking live view
Remove GoalsAssessmentComponent intermediary and lift its logic
directly into MarkingLive. Simplify the assessments grid to prompt
class selection before streaming data.
